[#DCM-146] Fix Eclipse warnings  * Fixed 'the expression of type int is boxed into Integer' warning

+++ dcm4che2/trunk/dcm4che-imageio-rle/src/main/java/org/dcm4che2/imageioimpl/plugins/rle/RLEImageReader.java
@@ -442,7 +442,8 @@ public class RLEImageReader extends ImageReader {
             }
         } catch (EOFException e) {
             log.warn("RLE Segment #{} too short, set missing {} bytes to 0",
-                    curSeg, (bs.length - pos + off) / pixelStride);
+                    Integer.valueOf(curSeg),
+                    Integer.valueOf((bs.length - pos + off) / pixelStride));
         }
     }
 
@@ -465,14 +466,14 @@ public class RLEImageReader extends ImageReader {
             }
         } catch (EOFException e) {
             log.warn("RLE Segment #{} too short, set missing {} bytes to 0",
-                    curSeg, bs.length - pos);
+                    Integer.valueOf(curSeg), Integer.valueOf(bs.length - pos));
         }
     }
 
     private int checkLengthTooLong(int length, int max) {
         if (length > max) {
             log.warn("RLE Segment #{} too long, truncate {} bytes",
-                    curSeg, length - max);
+                    Integer.valueOf(curSeg), Integer.valueOf(length - max));
             return max;
         }
         return length;
@@ -499,7 +500,7 @@ public class RLEImageReader extends ImageReader {
             }
         } catch (EOFException e) {
             log.warn("RLE Segment #{} too short, set missing {} bytes to 0",
-                    curSeg, ss.length - pos);
+                    Integer.valueOf(curSeg), Integer.valueOf(ss.length - pos));
         }
     }
